\name{exceptionClass}
\docType{class}

\alias{exception}
\alias{exceptionClass}
\alias{initialize, exceptionClass-method}
\alias{tryExc, exceptionClass-method}

\title{Class "exceptionClass". A class to represent a selenium exception}
\description{
  This class is designed to represent a selenium exception.
}
\details{
  The exception class is implemented as a reference class
  (see \code{\link{setRefClass}} for more details).
}

\section{Fields}{
  \describe{
     \item{\code{exceptionTable}:}{TODO}
     \item{\code{javaWarMes}:}{TODO}
  }
}

\section{Methods}{
  \describe{
     \item{\code{initialize}:}{TODO}
     \item{\code{tryExc}:}{TODO}
  }
}

\author{
  Aleix Ruiz de Villa, Andreu Vall, Lluis Ramon
}
\seealso{
  \code{\link{firefoxClass}}
}
