\name{javaKeysClass}
\docType{class}

\alias{javaKeys}
\alias{javaKeysClass}
\alias{initialize, javaKeysClass-method}

\title{Class "javaKeysClass". A class to represent a key}
\description{
  This class is designed to represent a key.
}
\details{
  The javaKeys class is implemented as a reference class
  (see \code{\link{setRefClass}} for more details).
}

\section{Fields}{
  \describe{
     \item{\code{keysArray}:}{TODO}
     \item{\code{keysNames}:}{TODO}
     \item{\code{keysOriginal}:}{TODO}
  }
}

\section{Methods}{
  \describe{
     \item{\code{initialize}:}{TODO}
  }
}

\author{
  Aleix Ruiz de Villa, Andreu Vall, Lluis Ramon
}
\seealso{
  \code{\link{firefoxClass}}
}
