\docType{package}
\name{relenium}
\alias{relenium}
\alias{relenium-package}
\title{relenium: powerful web scrapping}
\description{
    Selenium is a popular suite of tools specifically for automating web browsers. Selenium allows users to write tests in a number of popular programming languages including Java. The relenium package is intended to provide access to Selenium and use it for easy and powerful web scraping.
}

