\name{webElement}
\alias{initialize}
\alias{webElement}
\alias{webElement-method}
\title{Web Element class}
\usage{
  webElement(...)
}
\arguments{
  \item{...}{Additional argument list that might not ever
  used.}
}
\value{
  A webElement
}
\description{
  Some additional details about this S4 generic and its
  methods. The extra blank line between this section and
  the title is.
}
\author{
  Aleix Ruiz de Villa, Andreu Vall, Lluís Ramon
}
\seealso{
  \code{\link{firefox}}
}

