\name{remoteWebElementClass}
\docType{class}

\alias{remoteWebElement}
\alias{remoteWebElementClass}

\alias{clear, remoteWebElementClass-method} 
\alias{click, remoteWebElementClass-method} 
\alias{deselectAll, remoteWebElementClass-method} 
\alias{deselectByIndex, remoteWebElementClass-method} 
\alias{deselectByValue, remoteWebElementClass-method} 
\alias{deselectByVisibleText, remoteWebElementClass-method} 
\alias{findElementByClassName, remoteWebElementClass-method} 
\alias{findElementByCssSelector, remoteWebElementClass-method} 
\alias{findElementById, remoteWebElementClass-method} 
\alias{findElementByLinkText, remoteWebElementClass-method} 
\alias{findElementByName, remoteWebElementClass-method} 
\alias{findElementByPartialLinkText, remoteWebElementClass-method} 
\alias{findElementByTagName, remoteWebElementClass-method} 
\alias{findElementByXPath, remoteWebElementClass-method} 
\alias{findElementsByClassName, remoteWebElementClass-method} 
\alias{findElementsByCssSelector, remoteWebElementClass-method} 
\alias{findElementsById, remoteWebElementClass-method} 
\alias{findElementsByLinkText, remoteWebElementClass-method} 
\alias{findElementsByName, remoteWebElementClass-method} 
\alias{findElementsByPartialLinkText, remoteWebElementClass-method} 
\alias{findElementsByTagName, remoteWebElementClass-method} 
\alias{findElementsByXPath, remoteWebElementClass-method} 
\alias{getAllSelectedOptions, remoteWebElementClass-method} 
\alias{getAttribute, remoteWebElementClass-method} 
\alias{getCssValue, remoteWebElementClass-method} 
\alias{getFirstSelectedOption, remoteWebElementClass-method} 
\alias{getHtml, remoteWebElementClass-method} 
\alias{getId, remoteWebElementClass-method} 
\alias{getOptions, remoteWebElementClass-method} 
\alias{getSize, remoteWebElementClass-method} 
\alias{getTagName, remoteWebElementClass-method} 
\alias{getText, remoteWebElementClass-method} 
\alias{initialize, remoteWebElementClass-method} 
\alias{isDisplayed, remoteWebElementClass-method} 
\alias{isEnabled, remoteWebElementClass-method} 
\alias{isMultiple, remoteWebElementClass-method} 
\alias{isSelected, remoteWebElementClass-method} 
\alias{selectByIndex, remoteWebElementClass-method} 
\alias{selectByValue, remoteWebElementClass-method} 
\alias{selectByVisibleText, remoteWebElementClass-method} 
\alias{sendKeys, remoteWebElementClass-method} 
\alias{submit, remoteWebElementClass-method} 
\alias{keys, remoteWebElementClass-method} 
\alias{isMultiple, remoteWebElementClass-method} 
\alias{getOptions, remoteWebElementClass-method} 
\alias{deselectAll, remoteWebElementClass-method} 
\alias{getAllSelectedOptions, remoteWebElementClass-method} 
\alias{getFirstSelectedOption, remoteWebElementClass-method} 
\alias{selectByVisibleText, remoteWebElementClass-method} 
\alias{selectByIndex, remoteWebElementClass-method} 
\alias{selectByValue, remoteWebElementClass-method} 
\alias{deselectByValue, remoteWebElementClass-method} 
\alias{deselectByIndex, remoteWebElementClass-method} 
\alias{deselectByVisibleText, remoteWebElementClass-method} 

\title{Class "remoteWebElementClass". A class to represent a remote web element}
\description{
  This class is designed to represent a remote web element.
}
\details{
  The remoteWebElement class is implemented as a reference class
  (see \code{\link{setRefClass}} for more details).
}

\section{Fields}{
  \describe{
     \item{\code{javaKeys}:}{TODO}
     \item{\code{javaWebElement}:}{TODO}
     \item{\code{javaSelect}:}{TODO}
     \item{\code{keys}:}{TODO}
     \item{\code{methodNames}:}{TODO}
  }
}

\section{Methods}{
  \describe{
  
     \item{\code{clear}:}{TODO} 
     \item{\code{click}:}{TODO} 
     \item{\code{deselectAll}:}{TODO} 
     \item{\code{deselectByIndex}:}{TODO} 
     \item{\code{deselectByValue}:}{TODO} 
     \item{\code{deselectByVisibleText}:}{TODO} 
     \item{\code{findElementByClassName}:}{TODO} 
     \item{\code{findElementByCssSelector}:}{TODO} 
     \item{\code{findElementById}:}{TODO} 
     \item{\code{findElementByLinkText}:}{TODO} 
     \item{\code{findElementByName}:}{TODO} 
     \item{\code{findElementByPartialLinkText}:}{TODO} 
     \item{\code{findElementByTagName}:}{TODO} 
     \item{\code{findElementByXPath}:}{TODO} 
     \item{\code{findElementsByClassName}:}{TODO} 
     \item{\code{findElementsByCssSelector}:}{TODO} 
     \item{\code{findElementsById}:}{TODO} 
     \item{\code{findElementsByLinkText}:}{TODO} 
     \item{\code{findElementsByName}:}{TODO} 
     \item{\code{findElementsByPartialLinkText}:}{TODO} 
     \item{\code{findElementsByTagName}:}{TODO} 
     \item{\code{findElementsByXPath}:}{TODO} 
     \item{\code{getAllSelectedOptions}:}{TODO} 
     \item{\code{getAttribute}:}{TODO} 
     \item{\code{getCssValue}:}{TODO} 
     \item{\code{getFirstSelectedOption}:}{TODO} 
     \item{\code{getHtml}:}{TODO} 
     \item{\code{getId}:}{TODO} 
     \item{\code{getOptions}:}{TODO} 
     \item{\code{getSize}:}{TODO} 
     \item{\code{getTagName}:}{TODO} 
     \item{\code{getText}:}{TODO} 
     \item{\code{initialize}:}{TODO} 
     \item{\code{isDisplayed}:}{TODO} 
     \item{\code{isEnabled}:}{TODO} 
     \item{\code{isMultiple}:}{TODO} 
     \item{\code{isSelected}:}{TODO} 
     \item{\code{selectByIndex}:}{TODO} 
     \item{\code{selectByValue}:}{TODO} 
     \item{\code{selectByVisibleText}:}{TODO} 
     \item{\code{sendKeys}:}{TODO} 
     \item{\code{submit}:}{TODO} 
     \item{\code{keys}:}{TODO} 
     \item{\code{isMultiple}:}{TODO} 
     \item{\code{getOptions}:}{TODO} 
     \item{\code{deselectAll}:}{TODO} 
     \item{\code{getAllSelectedOptions}:}{TODO} 
     \item{\code{getFirstSelectedOption}:}{TODO} 
     \item{\code{selectByVisibleText}:}{TODO} 
     \item{\code{selectByIndex}:}{TODO} 
     \item{\code{selectByValue}:}{TODO} 
     \item{\code{deselectByValue}:}{TODO} 
     \item{\code{deselectByIndex}:}{TODO} 
     \item{\code{deselectByVisibleText}:}{TODO} 
     
  } 
}

\author{
  Aleix Ruiz de Villa, Andreu Vall, Lluis Ramon
}
\seealso{
  \code{\link{firefoxClass}}
}

