\name{remoteWebElementClass}
\docType{class}

\alias{remoteWebElement}
\alias{remoteWebElementClass}
\alias{initialize, remoteWebElementClass-method}
\alias{clear, remoteWebElementClass-method}

\title{Class "remoteWebElementClass". A class to represent a remote web element}
\description{
  This class is designed to represent a remote web element.
}
\details{
  The remoteWebElement class is implemented as a reference class
  (see \code{\link{setRefClass}} for more details).
}

\section{Fields}{
  \describe{
     \item{\code{javaKeys}:}{TODO}
     \item{\code{javaWebElement}:}{TODO}
     \item{\code{javaSelect}:}{TODO}
     \item{\code{keys}:}{TODO}
     \item{\code{methodNames}:}{TODO}
  }
}

\section{Methods}{
  \describe{
     \item{\code{initialize}:}{TODO}
     \item{\code{clear}:}{TODO}
  }
}

\author{
  Aleix Ruiz de Villa, Andreu Vall, Lluis Ramon
}
\seealso{
  \code{\link{firefoxClass}}
}

